# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: wesal
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://app.wesal.plus'
paths:
...
  '/api/messenger/{userId}/get-user-messages':
    get:
      summary: 'Get User Conversation'
      operationId: getUserConversation
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  response_token: 0
                  reaction: 2
                  incident: null
                  client_models: []
                  data:
                    message: 'The payload is invalid.'
                    auth_info:
                      authorized: false
                      reaction_code: 9
                properties:
                  response_token:
                    type: integer
                    example: 0
                  reaction:
                    type: integer
                    example: 2
                  incident:
                    type: string
                    example: null
                  client_models:
                    type: array
                    example: []
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'The payload is invalid.'
                      auth_info:
                        type: object
                        properties:
                          authorized:
                            type: boolean
                            example: false
                          reaction_code:
                            type: integer
                            example: 9
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: userId
        description: ''
        example: qui
        required: true
        schema:
          type: string
...